+++ src/main/java/com/salesforce/dataloader/ui/OAuthFlow.java
@@ -114,23 +114,4 @@ public abstract class OAuthFlow extends Dialog {
         new URIBuilder(url).getQueryParams().stream().forEach(kvp -> params.put(kvp.getName(), kvp.getValue()));
         return params;
     }
-    
-    public static void processSuccessfulLogin(InputStream httpResponseInputStream, Config config) throws IOException {
-
-        StringBuilder builder = new StringBuilder();
-        BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(httpResponseInputStream, "UTF-8"));
-        for (int c = bufferedReader.read(); c != -1; c = bufferedReader.read()) {
-            builder.append((char) c);
-        }
-
-        String jsonTokenResult = builder.toString();
-        Gson gson = new GsonBuilder()
-                .setFieldNamingPolicy(FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES)
-                .create();
-        OAuthToken token = gson.fromJson(jsonTokenResult, OAuthToken.class);
-        config.setValue(Config.OAUTH_ACCESSTOKEN, token.getAccessToken());
-        config.setValue(Config.OAUTH_REFRESHTOKEN, token.getRefreshToken());
-        config.setValue(Config.ENDPOINT, token.getInstanceUrl());
-    }
-
 }
